+++ WolvenKit.App/ViewModels/Editor/Basic/PropertiesViewModel.cs
@@ -1,9 +1,16 @@
+using System;
+using System.Collections.Generic;
+using System.ComponentModel;
+using System.Diagnostics;
+using System.IO;
+using System.Windows.Controls;
 using Catel;
 using Catel.Services;
+using NAudio.Wave;
 using WolvenKit.Common.Services;
 using WolvenKit.Functionality.Services;
 using WolvenKit.Models;
-
+using WolvenKit.Views.Editor.AudioTool;
 
 namespace WolvenKit.ViewModels.Editor
 {
@@ -49,6 +56,10 @@ namespace WolvenKit.ViewModels.Editor
             SetupToolDefaults();
 
             SetToNullAndResetVisibility();
+
+
+            nAudioSimple = NAudioSimpleEngine.Instance;
+            NAudioSimpleEngine.Instance.PropertyChanged += NAudioEngine_PropertyChanged;
         }
 
 
@@ -146,6 +157,124 @@ namespace WolvenKit.ViewModels.Editor
         public const string ToolTitle = "Properties";
         #endregion
 
+
+        #region AudioPreview
+
+        /// <summary>
+        /// Working directory for audio preview.
+        /// </summary>
+        private const string wdir = "vgmstream\\AudioWorkingDir\\";
+
+        /// <summary>
+        /// audio file list for opuspak type files with multiple wems inside. (not visually implemented)
+        /// </summary>
+        public List<TextBlock> AudioFileList { get; set; } = new List<TextBlock>();
+
+        /// <summary>
+        /// add an audio item. currently instantly loads this.
+        /// </summary>
+        /// <param name="path"></param>
+        public void AddAudioItem(string path) => TempConvertToWemWav(path);
+
+        /// <summary>
+        /// current track channel position
+        /// </summary>
+        public TimeSpan ChannelPosition { get; set; }
+
+        /// <summary>
+        /// Naudio engine
+        /// </summary>
+        public NAudioSimpleEngine nAudioSimple { get; set; }
+
+        /// <summary>
+        /// current track name.
+        /// </summary>
+        public string CurrentTrackName { get;set; }
+
+
+        /// <summary>
+        /// convert a file to wav to preview it.
+        /// </summary>
+        /// <param name="path"></param>
+        public void TempConvertToWemWav(string path)
+        {
+            string WKitAppData = Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData), "REDModding", "WolvenKit");
+
+            string ManagerCacheDir = Path.Combine(WKitAppData, "Temp_Audio");
+
+            //Clean directory
+            Directory.CreateDirectory(ManagerCacheDir);
+
+            foreach (var f in Directory.GetFiles(ManagerCacheDir))
+            {
+                try
+                {
+
+                        File.Delete(f);
+
+
+                }
+                catch
+                {
+
+                }
+            }
+
+            var outf = Path.Combine(ManagerCacheDir, Path.GetFileNameWithoutExtension(path) + ".wav");
+            Trace.WriteLine(outf);
+            Trace.WriteLine(path);
+
+            var arg = path + " -o " + outf;
+            var si = new ProcessStartInfo(
+                    AppDomain.CurrentDomain.BaseDirectory + "\\vgmstream\\test.exe",
+                    arg
+                )
+            {
+                CreateNoWindow = true,
+                WindowStyle = ProcessWindowStyle.Hidden,
+                UseShellExecute = false,
+                RedirectStandardError = true,
+                RedirectStandardOutput = true,
+                Verb = "runas"
+            };
+            var proc = Process.Start(si);
+            proc.WaitForExit();
+            Trace.WriteLine(proc.StandardOutput.ReadToEnd());
+
+
+            var lvi = new TextBlock()
+            {
+                Text = Path.GetFullPath(outf),
+                Tag = Path.GetFileName(outf)
+            };
+
+            NAudioSimpleEngine.Instance.OpenFile(outf);
+          CurrentTrackName   =Path.GetFileNameWithoutExtension(outf);
+
+            //AudioFileList.Add(lvi);
+        }
+
+        /// <summary>
+        /// property changed for naudio
+        /// </summary>
+        /// <param name="sender"></param>
+        /// <param name="e"></param>
+        private void NAudioEngine_PropertyChanged(object sender, PropertyChangedEventArgs e)
+        {
+            switch (e.PropertyName)
+            {
+
+
+                case "ChannelPosition":
+                    ChannelPosition = TimeSpan.FromSeconds(NAudioSimpleEngine.Instance.ChannelPosition);
+                    break;
+
+                default:
+                    // Do Nothing
+                    break;
+            }
+        }
+        #endregion
 
     }
 }
